Use int rather than short for bool; it will give more efficient code.

+++ src/include/curses.h
@@ -22,7 +22,7 @@
 #undef  TRUE
 #define TRUE	(1)
 
-typedef short bool;
+typedef int bool;
 typedef uint32_t chtype;
 typedef uint32_t attr_t;
